add hints
