+++ src/utils/hints/microblocksHints.js
@@ -1,0 +1,4 @@
+export const microblocksHints = {
+  transactionHash: 'Unique identifier of the transaction.',
+  transactionTimeime: 'Date and time when the transaction was created.',
+}
